+++ group.go
@@ -8,7 +8,7 @@ import (
 type Group struct {
 	prefix      string
 	middlewares []MiddlewareFunc
-	router      *Router
+	raptor      *Raptor
 }
 
 // Group :
@@ -16,7 +16,7 @@ func (g *Group) Group(prefix string, middleware ...MiddlewareFunc) *Group {
 	return &Group{
 		prefix:      fmt.Sprintf("%s%s", g.prefix, prefix),
 		middlewares: append(g.middlewares, middleware...),
-		router:      g.router,
+		raptor:      g.raptor,
 	}
 }
 
@@ -28,7 +28,7 @@ func (g *Group) addRoute(method string, path string, h HandlerFunc, middleware .
 	m := make([]MiddlewareFunc, 0, len(g.middlewares)+len(middleware))
 	m = append(m, g.middlewares...)
 	m = append(m, middleware...)
-	g.router.addRoute(method, fullPath, h, m...)
+	g.raptor.addRoute(method, fullPath, h, m...)
 	return g
 }